requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
Django==5.1.7
django-crispy-forms==2.3
frozenlist==1.5.0
idna==3.10
multidict==6.2.0
openai==0.28.1
pillow==11.1.0
propcache==0.3.1
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
yarl==1.18.3
